pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compass-ecl-mas"
version = "0.2.0"
description = "Constraint-based ethical governance layer (ECL) and Moral Alignment Score (MAS) for public-sector education decision support, with Pareto evaluation and reproducible artifacts."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Tales Matos"}]
keywords = ["responsible-ai", "governance", "fairness", "education", "pareto", "public-sector", "reproducibility"]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "reportlab>=4.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.32"]
data = ["ucimlrepo>=0.0.7"]
dev = ["pytest>=7.4", "ruff>=0.5.0"]
fairlearn = ["fairlearn==0.13.0", "scipy>=1.9.3,<1.16.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
